# openapi/skill-attestor.openapi.yaml
openapi: 3.1.0
info:
  title: Skill Attestor API
  version: 0.1.0
  description: Scan skill bundles and issue signed attestations.
servers:
  - url: http://localhost:8080
paths:
  /v1/scan:
    post:
      summary: Create a scan
      description: |
        Paid endpoint (x402).

        Returns 200 with an immediate scan result, or 202 with a scan ticket.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanCreateRequest'
      responses:
        '200':
          description: Scan completed synchronously.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResult'
        '202':
          description: Scan accepted for async processing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '402':
          $ref: '#/components/responses/PaymentRequired'
  /v1/scan/{scan_id}:
    get:
      summary: Get scan status or result
      parameters:
        - name: scan_id
          in: path
          required: true
          schema: { type: string }
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Scan record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResultOrStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/attest:
    post:
      summary: Issue an attestation for a completed scan
      description: Paid endpoint (x402).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttestRequest'
      responses:
        '200':
          description: Attestation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttestationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/attest/{root_sha256}:
    get:
      summary: Get attestations for an artifact root hash
      parameters:
        - name: root_sha256
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Attestations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttestationListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/verify:
    post:
      summary: Verify an attestation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
      responses:
        '200':
          description: Verification result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/policies/{policy_version}:
    get:
      summary: Get policy metadata
      parameters:
        - name: policy_version
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Policy metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/revocations:
    get:
      summary: Get revocation list
      responses:
        '200':
          description: Revocations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevocationsResponse'
components:
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PaymentRequired:
      description: Payment required (x402).
      headers:
        PAYMENT-REQUIRED:
          description: Base64-encoded JSON payment requirements.
          schema: { type: string }
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ScanCreateRequest:
      type: object
      additionalProperties: false
      properties:
        policy_version:
          type: string
          description: Optional; defaults to server default policy.
        share_anonymized_artifact_for_research:
          type: boolean
          default: false
        inline:
          $ref: '#/components/schemas/InlineInput'
        zip:
          $ref: '#/components/schemas/ZipInput'
        git:
          $ref: '#/components/schemas/GitInput'
      oneOf:
        - required: [inline]
        - required: [zip]
        - required: [git]
    InlineInput:
      type: object
      additionalProperties: false
      properties:
        filename: { type: string, default: SKILL.md }
        content: { type: string, description: UTF-8 text content }
      required: [content]
    ZipInput:
      type: object
      additionalProperties: false
      properties:
        zip_b64: { type: string, description: Base64-encoded zip bytes }
        entrypoint: { type: string, default: SKILL.md }
      required: [zip_b64]
    GitInput:
      type: object
      additionalProperties: false
      properties:
        repo_url: { type: string }
        ref: { type: string }
        entrypoint: { type: string, default: SKILL.md }
      required: [repo_url, ref]

    ScanAccepted:
      type: object
      additionalProperties: false
      properties:
        scan_id: { type: string }
        status: { type: string, enum: [queued, running] }
        result_token: { type: string }
      required: [scan_id, status, result_token]

    ScanResultOrStatus:
      oneOf:
        - $ref: '#/components/schemas/ScanStatus'
        - $ref: '#/components/schemas/ScanResult'

    ScanStatus:
      type: object
      additionalProperties: false
      properties:
        scan_id: { type: string }
        status: { type: string, enum: [queued, running, failed] }
        created_at: { type: string }
        updated_at: { type: string }
        error:
          $ref: '#/components/schemas/ErrorInfo'
      required: [scan_id, status, created_at, updated_at]

    ScanResult:
      type: object
      additionalProperties: false
      properties:
        scan_id: { type: string }
        status: { type: string, enum: [completed] }
        created_at: { type: string }
        updated_at: { type: string }
        policy_version: { type: string }
        artifact_root_sha256: { type: string }
        verdict: { type: string, enum: [pass, warn, fail] }
        score: { type: integer }
        finding_counts:
          $ref: '#/components/schemas/FindingCounts'
        result_token:
          type: string
          description: Capability token for future status or attestation retrieval.
        findings:
          type: array
          items:
            $ref: '#/components/schemas/Finding'
      required:
        - scan_id
        - status
        - created_at
        - updated_at
        - policy_version
        - artifact_root_sha256
        - verdict
        - score
        - finding_counts
        - findings

    FindingCounts:
      type: object
      additionalProperties: false
      properties:
        critical: { type: integer }
        high: { type: integer }
        medium: { type: integer }
        low: { type: integer }
      required: [critical, high, medium, low]

    Finding:
      type: object
      additionalProperties: false
      properties:
        finding_id: { type: string }
        severity: { type: string, enum: [critical, high, medium, low] }
        category: { type: string }
        title: { type: string }
        description: { type: string }
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/Evidence'
        recommendations:
          type: array
          items: { type: string }
        fingerprints:
          type: array
          items: { type: string }
      required:
        - finding_id
        - severity
        - category
        - title
        - description
        - evidence
        - recommendations
        - fingerprints

    Evidence:
      type: object
      additionalProperties: false
      properties:
        path: { type: string }
        line_start: { type: integer }
        line_end: { type: integer }
        snippet: { type: string }
      required: [path, line_start, line_end, snippet]

    AttestRequest:
      type: object
      additionalProperties: false
      properties:
        scan_id: { type: string }
        token: { type: string }
      required: [scan_id, token]

    AttestationResponse:
      type: object
      additionalProperties: false
      properties:
        attestation_document: { type: object }
        signature_b64url: { type: string }
      required: [attestation_document, signature_b64url]

    AttestationListResponse:
      type: object
      additionalProperties: false
      properties:
        artifact_root_sha256: { type: string }
        attestations:
          type: array
          items:
            $ref: '#/components/schemas/AttestationResponse'
      required: [artifact_root_sha256, attestations]

    VerifyRequest:
      type: object
      additionalProperties: false
      properties:
        attestation_document: { type: object }
        signature_b64url: { type: string }
        artifact_manifest:
          description: Optional; if provided, recompute root hash.
          type: array
          items:
            $ref: '#/components/schemas/ManifestEntry'
      required: [attestation_document, signature_b64url]

    VerifyResponse:
      type: object
      additionalProperties: false
      properties:
        valid_signature: { type: boolean }
        matches_artifact_hash: { type: boolean }
        revoked: { type: boolean }
        errors:
          type: array
          items: { type: string }
      required: [valid_signature, matches_artifact_hash, revoked, errors]

    ManifestEntry:
      type: object
      additionalProperties: false
      properties:
        path: { type: string }
        sha256: { type: string }
        size_bytes: { type: integer }
      required: [path, sha256, size_bytes]

    PolicyResponse:
      type: object
      additionalProperties: false
      properties:
        policy_version: { type: string }
        verdict_thresholds: { type: object }
        rules: { type: object }
      required: [policy_version, verdict_thresholds, rules]

    RevocationsResponse:
      type: object
      additionalProperties: false
      properties:
        revocations_document: { type: object }
        signature_b64url: { type: string }
      required: [revocations_document, signature_b64url]

    ErrorInfo:
      type: object
      additionalProperties: false
      properties:
        error_code: { type: string }
        message: { type: string }
      required: [error_code, message]

    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        error:
          $ref: '#/components/schemas/ErrorInfo'
      required: [error]